+++ R/spida2.R
@@ -6,7 +6,7 @@
 #'
 #' @section Wald tests and linear hypothesis matrices:
 #' \itemize{
-#' \item \code{\link{wald}} quick Wald tests -- optionally building hypothesis matrices from regular expressions -- with full or reduced rank
+#' \item \code{\link{wald}} Wald tests with L matrices optionally created with regular expressions. Uses SVD to handle linear dependencies in rows of L
 #' \item \code{\link{walddf}} version of wald that returns a data frame
 #' \item \code{\link{as.data.frame.wald}} return a data frame from a wald object
 #' \item \code{\link{coef.wald}} method to extract estimated coefficients
@@ -28,22 +28,22 @@
 #' }
 #' @section Multilevel data frames:
 #' \itemize{
-#' \item \code{\link{tolong}} and \code{\link{towide}} are 
-#'   interfaces to \code{\link{stats::reshape}} to facilitate
-#'   the typical uses of reshape for longitudinal data.
 #' \item \code{\link{capply}}: \code{capply(x,id,FUN)} applies the
 #'   function \code{FUN} to chunks of 'x' formed by levels of 'id'.
 #'   The result has the same form as 'x' with replication within
 #'   chunks, if needed.
-#' \item \code{\link{cvar}} and \code{\link{dvar}} are designed to be used
-#'   in linear model formulas to generate 'centered-within-group' and 'within-group
-#'   deviation' variables. WIth factors, they generate mean incidence matrices.
 #' \item \code{\link{up}}, \code{\link{agg}} and \code{\link{up_apply}} create summary
 #'   data sets consisting, by default, of within-id-invariant variables. 
 #'   Summaries of id-varying variables can also be included. \code{\link{agg}} can
 #'   create mean incidence matrices for lower-level factors.
-#' \item \code{link{varLevel}}: the level of a variable with respect
-#'   to a clustering formula.
+#' \item \code{\link{cvar}} and \code{\link{dvar}} are designed to be used
+#'   in linear model formulas to generate 'centered-within-group' and 'within-group
+#'   deviation' variables. WIth factors, they generate mean incidence matrices.
+#' \item \code{link{varLevel}} and \item \code{link{gicc}}: the level of a variable with respect
+#'   to a clustering formula and the 'generalized' intra-class correlation coefficient.
+#' \item \code{\link{tolong}} and \code{\link{towide}} are 
+#'   interfaces to \code{\link{stats::reshape}} to facilitate
+#'   the typical uses of reshape for longitudinal data.
 #' }
 #' @section Splines -- parametric and non-parametric:
 #' \itemize{
@@ -65,10 +65,16 @@
 #' \item \code{\link{coffee}} Artifical data on coffee, heart damage and stress. Illustrates Simpson's Paradox with continuous predictors.
 #' \item \code{\link{hw}} Artifical data on height, weight and health. Illustrates suppression.
 #' }
+#' @section Graphics
+#' \itemize {
+#' \item \code{\link{gd}} and \code{\link{td}} easy interface to set
+#'       graphical parameters for lattice and graphics. \code{gd} sets
+#'       parameters to make graphs look like ggplot2 graphics. 
+#' \item \code{\link{panel.fit}} add fitted values and error bands with \code{\link{latticeExtra::layer}} or \code{\link{latticeExtra::glayer}}
+#' \item \code{\link{panel.dell}} add data ellipse \code{\link{latticeExtra::layer}} or \code{\link{latticeExtra::glayer}}
+#' }
 #' @section Miscellaneous utility functions:
 #' \itemize{
-#' \item \code{\link{gd}} and \code{\link{td}} easy interface to set
-#'       graphical parameters for lattice graphsics.
 #' \item \code{\link{sortdf}} sort rows of a data frame -- useful in a magrittr pipeline
 #' \item \code{\link{assn}} assign -- useful in a magrittr pipeline
 #' \item \code{\link{disp}} utility to display value of a variable -- useful for debugging
@@ -83,7 +89,7 @@
 #' \item \code{\link{run}} evaluate a string as a command with try
 #' \item \code{\link{grepv}} grep(..., value = TRUE)
 #' }
-#' @section String manipulatio functions:
+#' @section String manipulation functions:
 #' Tools to help manipulate messy data designed to work smoothly with \code{magrittr} pipes. 
 #' \itemize{
 #' \item \code{\link{sub_}} and \code{\link{gsub_}} handle substitution in a pipeline and return a factor if the input is a factor