+++ R/wald.R
@@ -66,21 +66,12 @@
 #' There are number of functions to help construct hypothesis matrices. See in
 #' particular \code{\link{Lfx}}.
 #'
-#' # To extend the 'wald' function to a new class of objects, one needs to
-#' write a 'getFix' # method to extract estimated coefficients, their estimated
-#' covariance matrix, and the # denominator degrees of freedom for each
-#' estimated coefficient. The getFix method for # glm objects is:
-#'
-#' getFix.glm <- function(fit,...) { ss <- summary(fit) ret <- list() ret$fixed
-#' <- coef(fit) ret$vcov <- vcov(fit) ret$df <- rep(ss$df.residual,
-#' length(ret$fixed)) ret }
-#'
-#' # and for 'mipo' objects in the packages 'mice':
-#'
-#' getFix.mipo <- function( fit, ...){ # # pooled multiple imputation object in
-#' mice # 'wald' will use the minimal df for components with non-zero weights #
-#' -- this is probably too conservative and should be improved # ret <- list()
-#' ret$fixed <- fit$qbar ret$vcov <- fit$t ret$df <- fit$df ret }
+#' To extend the 'wald' function to a new class of objects, one needs to
+#' write a 'getFix' method to extract estimated coefficients, their estimated
+#' covariance matrix, and the denominator degrees of freedom for each
+#' estimated coefficient. See the examples below for 
+#' a \code{\link{gefFix}} method for a \code{\link{glm}} object
+#' and for 'mipo' objects in the packages 'mice':
 #'
 #' @param fit a model for which a \code{getFix} method exists.
 #' @param Llist a hypothesis matrix or a pattern to be matched or a list of
@@ -108,6 +99,8 @@
 #' @param method 'svd' (current default) or 'qr' is the method used to find the
 #'       full rank version of the hypothesis matrix.  'svd' has correctly identified
 #'       the rank of a large hypothesis matrix where 'qr' has failed.
+#' @param pars passed to \code{\link{rstan::extract}} method for stanfit objects.
+#' @param include passed to \code{\link{rstan::extract}} method for stanfit objects.#' 
 #' @param help obsolete
 #' @return An object of class \code{wald}, with the following components:
 #'       COMPLETE
@@ -157,7 +150,30 @@
 #' require(lattice)
 #' xyplot( coef + U2 + L2 ~ ses | Sex, wald.dd,
 #'  main= 'Increase in predicted mathach per unit increase in ses')
-#'
+#' 
+#' # Example of a getFix method for a glm oject:
+#' 
+#' getFix.glm <- function(fit,...) { 
+#'   ss <- summary(fit) 
+#'   ret <- list() 
+#'   ret$fixed <- coef(fit) 
+#'   ret$vcov <- vcov(fit) 
+#'   ret$df <- rep(ss$df.residual,length(ret$fixed)) 
+#'   ret
+#' } 
+#' 
+#' # Example of a getFix method for a mipo object in the mice package:
+#' 
+#' getFix.mipo <- function( fit, ...){ 
+#'   # pooled multiple imputation object in mice 
+#'   # 'wald' will use the minimal df for components with non-zero weights 
+#'   #  -- this is probably too conservative and should be improved 
+#'   ret <- list()
+#'   ret$fixed <- fit$qbar 
+#'   ret$vcov <- fit$t 
+#'   ret$df <- fit$df 
+#'   ret 
+#' }
 #' @export
 wald <- function(fit, Llist = "", clevel = 0.95,
                  pred = NULL,
@@ -339,6 +355,188 @@ wald <- function(fit, Llist = "", clevel = 0.95,
   attr(ret,"class") <- "wald"
   ret
 }
+# New version with support for stanfit
+wald <- 
+  function(fit, Llist = "", clevel = 0.95,
+           pred = NULL,
+           data = NULL, debug = FALSE , maxrows = 25,
+           full = FALSE, fixed = FALSE,
+           invert = FALSE, method = 'svd',
+           df = NULL, pars = NULL,...) {
+    if (full) return(wald(fit, getX(fit)))
+    if(!is.null(pred)) return(wald(fit, getX(fit,pred)))
+    dataf <- function(x,...) {
+      x <- cbind(x)
+      rn <- rownames(x)
+      if(length(unique(rn)) < length(rn)) rownames(x) <- NULL
+      data.frame(x, ...)
+    }
+    as.dataf <- function(x, ...) {
+      x <- cbind(x)
+      rn <- rownames(x)
+      if(length(unique(rn)) < length(rn)) rownames(x) <- NULL
+      as.data.frame(x, ...)
+    }
+    unique.rownames <- function(x) {
+      ret <- c(tapply(1:length(x), x, function(xx) {
+        if(length(xx) == 1) ""
+        else 1:length(xx)
+      })) [tapply(1:length(x), x)]
+      ret <- paste(x, ret, sep="")
+      ret
+    }
+    if(is.character(Llist) ) Llist <- structure(list(Llist), names=Llist)
+    if(!is.list(Llist)) Llist <- list(Llist)
+    
+    ret <- list()
+    fix <- if(is.null(pars)) getFix(fit) else getFix(fit,pars=pars,...)
+    beta <- fix$fixed
+    vc <- fix$vcov
+    
+    dfs <- if(is.null(df) ) fix$df else df + 0*fix$df
+    for (ii in 1:length(Llist)) {
+      ret[[ii]] <- list()
+      Larg <- Llist[[ii]]
+      # Create hypothesis matrix: L
+      L <- NULL
+      if(is.character(Larg)) {
+        L <- Lmat(fit,Larg, fixed = fixed, invert = invert)
+      } else {
+        if(is.numeric(Larg)) {   # indices for coefficients to test
+          if(is.null(dim(Larg))) {
+            if(debug) disp(dim(Larg))
+            if((length(Larg) < length(beta)) && (all(Larg>0)||all(Larg<0)) ) {
+              L <- diag(length(beta))[Larg,]
+              dimnames(L) <- list( names(beta)[Larg], names(beta))
+            } else L <- rbind( Larg )
+          }
+          else L <- Larg
+        }
+      }
+      if (debug) {
+        disp(Larg)
+        disp(L)
+      }
+      # get data attribute, if any, in case it gets dropped
+      Ldata <- attr( L , 'data')
+      
+      ## identify rows of L that are not estimable because they depend on betas that are NA
+      Lna <- L[, is.na(beta), drop = FALSE]
+      narows <- apply(Lna,1, function(x) sum(abs(x))) > 0
+      
+      L <- L[, !is.na(beta),drop = FALSE]
+      ## restore the data attribute
+      attr(L,'data') <- Ldata
+      beta <- beta[ !is.na(beta) ]
+      
+      ## Anova
+      if( method == 'qr' ) {
+        qqr <- qr(t(na.omit(L)))
+        # Qqr <- Q(t(L))
+        L.rank <- qqr$rank
+        # L.rank <- attr(Qqr,'rank')
+        # L.miss <- attr(Qqr,'miss')
+        if(debug)disp( t( qr.Q(qqr)))
+        L.full <- t(qr.Q(qqr))[ 1:L.rank,,drop=FALSE]
+        #L.full <- t(Qqr[!L.miss,])[ 1:L.rank,,drop=F]
+      } else if ( method == 'svd' ) {
+        if(debug) disp(L)
+        #              if(debug)disp( t(na.omit(t(L))))
+        #              sv <- svd( t(na.omit(t(L))) , nu = 0 )
+        sv <- svd( na.omit(L) , nu = 0 )
+        
+        if(debug)disp( sv )
+        tol.fac <- max( dim(L) ) * max( sv$d )
+        if(debug)disp( tol.fac )
+        if ( tol.fac > 1e6 ) warning( "Poorly conditioned L matrix, calculated numDF may be incorrect")
+        tol <- tol.fac * .Machine$double.eps
+        if(debug)disp( tol )
+        L.rank <- sum( sv$d > tol )
+        if(debug)disp( L.rank )
+        if(debug)disp( t(sv$v))
+        L.full <- t(sv$v)[seq_len(L.rank),,drop = FALSE]
+      } else stop("method not implemented: choose 'svd' or 'qr'")
+      
+      # from package(corpcor)
+      # Note that the definition tol= max(dim(m))*max(D)*.Machine$double.eps
+      # is exactly compatible with the conventions used in "Octave" or "Matlab".
+      
+      if (debug && method == "qr") {
+        disp(qqr)
+        disp(dim(L.full))
+        disp(dim(vc))
+        disp(vc)
+      }
+      if (debug) disp(L.full)
+      if (debug) disp(vc)
+      
+      vv <-  L.full %*% vc %*% t(L.full)
+      eta.hat <- L.full %*% beta
+      Fstat <- (t(eta.hat) %*% qr.solve(vv,eta.hat,tol=1e-10)) / L.rank
+      included.effects <- apply(L,2,function(x) sum(abs(x),na.rm=TRUE)) != 0
+      denDF <- min( dfs[included.effects])
+      numDF <- L.rank
+      ret[[ii]]$anova <- list(numDF = numDF, denDF = denDF,
+                              "F-value" = Fstat,
+                              "p-value" = pf(Fstat, numDF, denDF, lower.tail = FALSE))
+      ## Estimate
+      
+      etahat <- L %*% beta
+      
+      # NAs if not estimable:
+      
+      etahat[narows] <- NA
+      if( nrow(L) <= maxrows ) {
+        etavar <- L %*% vc %*% t(L)
+        etasd <- sqrt( diag( etavar ))
+      } else {
+        etavar <- NULL
+        etasd <- sqrt( apply( L * (L%*%vc), 1, sum))
+      }
+      
+      denDF <- apply( L , 1 , function(x,dfs) min( dfs[x!=0]), dfs = dfs)
+      
+      aod <- cbind(
+        Estimate=c(etahat),
+        Std.Error = etasd,
+        DF = denDF,
+        "t-value" = c(etahat/etasd),
+        "p-value" = 2*pt(abs(etahat/etasd), denDF, lower.tail =FALSE))
+      colnames(aod)[ncol(aod)] <- 'p-value'
+      if (debug ) disp(aod)
+      if ( !is.null(clevel) ) {
+        #print(aod)
+        #print(aod[,'DF'])
+        #print(aod[,'etasd'])
+        hw <- qt(1 - (1-clevel)/2, aod[,'DF']) * aod[,'Std.Error']
+        #print(hw)
+        aod <- cbind( aod, LL = aod[,"Estimate"] - hw, UL = aod[,"Estimate"] + hw)
+        #print(aod)
+        if (debug ) disp(colnames(aod))
+        labs <- paste(c("Lower","Upper"), format(clevel))
+        colnames(aod) [ ncol(aod) + c(-1,0)] <- labs
+      }
+      if (debug ) disp(rownames(aod))
+      aod <- as.dataf(aod)
+      
+      rownames(aod) <- rownames(as.dataf(L))
+      labs(aod) <- names(dimnames(L))[1]
+      ret[[ii]]$estimate <- aod
+      ret[[ii]]$coef <- c(etahat)
+      ret[[ii]]$vcov <- etavar
+      ret[[ii]]$L <- L
+      ret[[ii]]$se <- etasd
+      ret[[ii]]$L.full <- L.full
+      ret[[ii]]$L.rank <- L.rank
+      if( debug ) disp(attr(Larg,'data'))
+      data.attr <- attr(Larg,'data')
+      if(is.null(data.attr) && !(is.null(data))) data.attr <- data
+      ret[[ii]]$data <- data.attr
+    }
+    names(ret) <- names(Llist)
+    attr(ret,"class") <- "wald"
+    ret
+  }
 
 # Test
 if(FALSE){
@@ -953,6 +1151,18 @@ getFix.MCMCglmm <- function(fit,...) {
   ret <- list()
   ret$fixed <- apply(fit$Sol, 2, mean)
   ret$vcov <- var( fit $ Sol)
+  ret$df <- rep(Inf, length(ret$fixed))
+  ret
+}
+#' @describeIn getFix method for `stanfit` objects in the `rstan` package
+#' @export
+getFix.stanfit <-
+function(fit, pars, include = TRUE, ...) {
+  if(missing(pars)) pars <- dimnames(fit)$parameter
+  sam <- as.matrix(fit, pars = pars , include = include)
+  ret <- list()
+  ret$fixed <- apply(sam, 2, mean)
+  ret$vcov <- var(sam)
   ret$df <- rep(Inf, length(ret$fixed))
   ret
 }